refactor(cli): split args.rs into 6 focused sub-modules with #[command(flatten)]
