+++ crates/webfang_core/src/cli/args/ai.rs
@@ -1,0 +1,28 @@
+use clap::Args;
+
+#[derive(Args, Debug, Default)]
+pub struct AiArgs {
+    /// Relevance threshold for AI semantic filtering (0.0-1.0)
+    #[cfg(feature = "ai")]
+    #[arg(long, default_value = "0.3", env = "WEBFANG_THRESHOLD")]
+    #[clap(next_help_heading = "AI Settings")]
+    pub threshold: f32,
+
+    /// Maximum tokens per chunk for AI processing
+    #[cfg(feature = "ai")]
+    #[arg(long, default_value = "32768", env = "WEBFANG_MAX_TOKENS")]
+    #[clap(next_help_heading = "AI Settings")]
+    pub max_tokens: usize,
+
+    /// Run AI model in offline mode
+    #[cfg(feature = "ai")]
+    #[arg(long, default_value = "false", env = "WEBFANG_OFFLINE", action = clap::ArgAction::SetTrue)]
+    #[clap(next_help_heading = "AI Settings")]
+    pub offline: bool,
+
+    /// AI model to use: granite-97m (default, fast) or granite-311m (higher quality)
+    #[cfg(feature = "ai")]
+    #[arg(long, env = "AI_MODEL_ID", value_parser = ["granite-97m", "granite-311m"])]
+    #[clap(next_help_heading = "AI Settings")]
+    pub ai_model: Option<String>,
+}
